requirements.txt
aiohttp>=3.9
langchain-core>=0.3
langchain-openai>=0.3
markdown>=3.5
matrix-nio>=0.25
python-telegram-bot>=20.0
PyYAML>=6.0
requests>=2.32
